split paths out into its own file
